requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
Flask-Login
email-validator
